implement polyphonic pulsewave synth
This basically works, but the audio has some crackles and notes
stick or drop relatively often. I think the USB stack is fighting
with the audio stack.

Originally, I planned to use the 3 buttons on the Fruit Jam board
for volume adjustment and maybe to change the patch. But, removing
my button polling code noticably reduced the amount of stuck
notes. I also commented out the code for printing MIDI events on
the USB serial console because that helped a bit more (not as much
as removing the button code though).

Another issue: This code only plays on the left channel, and I have
no idea why. As far as I can tell, the DAC volume stuff is set up
for stereo. But, when I turn the gain up on my mixer, there's noise
on the left channel, but the right is totally silent. Possible
there might be some kind of wiring problem on my prototype board.
Or, very likely, I may have done something wrong with the DAC or
synthio setup.