refactor audio setup, default to low volume
This moves the audio and synth setup stuff to its own function and
sets the default volume to a safe (low) earbud volume. The point is
to make this a bit safer to share without risking people's ears.
I want proper volume control eventually, but for now this is fine.