updates for production Fruit Jam, DAC, etc
Since I first wrote the guide for this, the TLV320DAC library has
been updated for better volume control. Also, the rev D production
Fruit Jam boards are generally available in the shop, and the
latest CircuitPython beta version is 10.0.0-beta.3.

This commit updates the code to work better with those changes.